# specs/1-todo-feature-expansion/plan/contracts/task-api-contracts.yaml
openapi: 3.0.3
info:
  title: Todo App Task API
  description: API for managing tasks with advanced features
  version: 1.0.0
paths:
  /api/tasks:
    get:
      summary: Get user's tasks with filtering, sorting, and search
      parameters:
        - name: skip
          in: query
          description: Number of records to skip
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: Maximum number of records to return
          required: false
          schema:
            type: integer
            default: 100
        - name: status
          in: query
          description: Filter by completion status
          required: false
          schema:
            type: string
            enum: [all, active, completed]
            default: all
        - name: priority
          in: query
          description: Filter by priority level
          required: false
          schema:
            type: string
            enum: [high, medium, low]
        - name: tag
          in: query
          description: Filter by tag name
          required: false
          schema:
            type: string
        - name: due_after
          in: query
          description: Filter tasks due after this date
          required: false
          schema:
            type: string
            format: date
        - name: due_before
          in: query
          description: Filter tasks due before this date
          required: false
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [due_date, priority, title, created_at]
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          description: Search term for title and description
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /api/tasks/search:
    post:
      summary: Search tasks by keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                q:
                  type: string
                  description: Search query
                filters:
                  type: object
                  properties:
                    status:
                      type: string
                      enum: [all, active, completed]
                    priority:
                      type: string
                      enum: [high, medium, low]
                    tags:
                      type: array
                      items:
                        type: string
              required:
                - q
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /api/tasks/{task_id}:
    get:
      summary: Get a specific task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Task not found

  /api/tags:
    get:
      summary: Get user's tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '200':
          description: Created tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input

  /api/tasks/{task_id}/tags:
    put:
      summary: Update tags for a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag_ids:
                  type: array
                  items:
                    type: string
                  description: List of tag IDs to associate with the task
              required:
                - tag_ids
      responses:
        '200':
          description: Updated task with tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
          minLength: 1
        description:
          type: string
        completed:
          type: boolean
          default: false
        user_id:
          type: string
          format: uuid
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        due_date:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        recurrence_pattern:
          type: object
          properties:
            type:
              type: string
              enum: [daily, weekly, monthly, custom]
            interval:
              type: integer
            days:
              type: array
              items:
                type: string
                enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        original_task_id:
          type: string
          format: uuid
      required:
        - id
        - title
        - completed
        - user_id
        - version
        - created_at
        - updated_at

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        due_date:
          type: string
          format: date-time
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
        recurrence_pattern:
          type: object
          properties:
            type:
              type: string
              enum: [daily, weekly, monthly, custom]
            interval:
              type: integer
            days:
              type: array
              items:
                type: string
                enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
        description:
          type: string
        completed:
          type: boolean
        priority:
          type: string
          enum: [high, medium, low]
        due_date:
          type: string
          format: date-time
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
        version:
          type: integer
      required:
        - version

    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
          minLength: 1
        color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$'
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - user_id
        - created_at

    TagCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
        color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$'
      required:
        - name